+++ hsb-things/package.json
@@ -14,13 +14,14 @@
 		"prod": "NODE_ENV=production node -r source-map-support/register build/main.js"
 	},
 	"dependencies": {
+		"express": "^4.17.1",
+		"express-ws": "^4.0.0",
 		"hsb-service-utils": "0.0.0",
 		"lodash": "^4.17.15",
 		"node-fetch": "^2.6.0",
 		"node-tradfri-client": "^2.1.2",
 		"serialport": "^8.0.7",
-		"source-map-support": "^0.5.16",
-		"ws": "^7.2.1"
+		"source-map-support": "^0.5.16"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.7.5",
